move cask/cmd/reinstall to cask/reinstall
