+++ Library/Homebrew/cask/cmd.rb
@@ -12,7 +12,6 @@ require "cask/config"
 require "cask/cmd/abstract_command"
 require "cask/cmd/audit"
 require "cask/cmd/install"
-require "cask/cmd/reinstall"
 
 module Cask
   # Implementation of the `brew cask` command-line interface.
