+++ Library/Homebrew/test/cask/reinstall_spec.rb
@@ -1,7 +1,10 @@
 # typed: false
 # frozen_string_literal: true
 
-describe Cask::Cmd::Reinstall, :cask do
+require "cask/installer"
+require "cask/reinstall"
+
+describe Cask::Reinstall, :cask do
   it "displays the reinstallation progress" do
     caffeine = Cask::CaskLoader.load(cask_path("local-caffeine"))
 
@@ -20,7 +23,7 @@ describe Cask::Cmd::Reinstall, :cask do
     EOS
 
     expect do
-      described_class.run("local-caffeine")
+      described_class.reinstall_casks(Cask::CaskLoader.load("local-caffeine"))
     end.to output(output).to_stdout
   end
 
@@ -45,7 +48,7 @@ describe Cask::Cmd::Reinstall, :cask do
     EOS
 
     expect do
-      described_class.run("local-caffeine", "--zap")
+      described_class.reinstall_casks(Cask::CaskLoader.load("local-caffeine"), zap: true)
     end.to output(output).to_stdout
   end
 
@@ -54,14 +57,14 @@ describe Cask::Cmd::Reinstall, :cask do
 
     expect(Cask::CaskLoader.load(cask_path("local-transmission"))).to be_installed
 
-    described_class.run("local-transmission")
+    described_class.reinstall_casks(Cask::CaskLoader.load("local-transmission"))
     expect(Cask::CaskLoader.load(cask_path("local-transmission"))).to be_installed
   end
 
   it "allows reinstalling a non installed Cask" do
     expect(Cask::CaskLoader.load(cask_path("local-transmission"))).not_to be_installed
 
-    described_class.run("local-transmission")
+    described_class.reinstall_casks(Cask::CaskLoader.load("local-transmission"))
     expect(Cask::CaskLoader.load(cask_path("local-transmission"))).to be_installed
   end
 end
